“选择是否为指针基本变量还是自定义类型”

+++ include/stl/threadpool.h
@@ -40,13 +40,13 @@ namespace stl {
         pthread_t tid_2;
         //明天实现自定义vector
         std::vector<ThreadPoolAssisted> Assisted;
-        ThreadPool(int a) {
+        ThreadPool(size_t a) {
             // pthread_create(&tid_1, nullptr, runtime_init, this);
             // pthread_create(&tid_2, nullptr, runtime_init, this);
         };
 
         static void* runtime_init(void* arg) {
-            //在vector中取出任务
+            // 在vector中取出任务
             ThreadPool* pool = static_cast<ThreadPool*>(arg);
             for (int i = 0; i < (pool->Assisted).size(); i++) {
                 ThreadPoolAssisted& task = pool->Assisted[i];
